tests/wpt/tests/css/css-backgrounds/parsing/background-size-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-size with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-size">
<meta name="assert" content="background-size supports only the grammar '<bg-size>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Blink and WebKit fail these by accepting negative values.
test_invalid_value("background-size", "-1px");
test_invalid_value("background-size", "2% -3%");

test_invalid_value("background-size", "1px 2px 3px");
</script>
</body>
</html>
